- fix bug on storing long-path name data
Previous assumption was that there will always
be a *single* tar header block written for any value,
which is not true if path-names are longer than 100
characters (in which case multiple extended header
blocks are being written).

Fix bug by computing the effective header block
size by knowing that the payload will be stored as
a multiple of the blocksize nonetheless. This is
a transparent backwards-compatible fix to also
support storing zarr data with long path lengths.