Enable timestamps in GitLab CI logs
Timestamps will be printed for each log line of a job, which can help
find individual operations that are not performing as expected.

A small drawback: timestamps are part of the raw log files. This may
require some post-processing by users if they want to diff these files.

Jira NRE-1410